requirements.txt
flask
uvicorn
requests
pyserial
open-cvpython-headless